Global Go code refactoring
Reorganized the bot code into focused subpackages and rewired imports/tests to match, without altering behavior. The handlers now live under pkg/bot/handlers, game logic under pkg/bot/game, import/export under pkg/bot/importexport, reminders under pkg/bot/reminders, and shared test DB setup under pkg/internal/testutil/db.go, with cmd/tg-word-reminder/main.go updated to use the new entrypoints.

  - Split pkg/bot/handlers.go into pkg/bot/handlers/default.go, pkg/bot/handlers/start.go, pkg/bot/handlers/settings.go, pkg/bot/handlers/pairs.go, pkg/bot/handlers/
    export.go, pkg/bot/handlers/game.go.
  - Moved game logic into pkg/bot/game/session.go and pkg/bot/game/message.go, adding game.DefaultManager and game.ResetDefaultManager plus exported result structs for
    cross‑package use.
  - Moved CSV logic into pkg/bot/importexport/csv.go and reminders into pkg/bot/reminders/tickers.go, updating call sites to new package APIs.
  - Relocated tests to mirror packages and switched DB setup to pkg/internal/testutil/db.go, adjusting game handler tests to use the exported manager API.
  - Updated cmd/tg-word-reminder/main.go to import pkg/bot/handlers, pkg/bot/game, and pkg/bot/reminders.
