+++ pkg/bot/handlers/settings_test.go
@@ -1,4 +1,4 @@
-package bot
+package handlers
 
 import (
 	"errors"
